requirements.txt
atlassian-python-api==1.15.5
certifi==2019.11.28
chardet==3.0.4
idna==2.9
oauthlib==3.1.0
python-gitlab==2.0.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
urllib3==1.25.8
